SVG forced colors follow-up
Partial revert of:
https://chromium-review.googlesource.com/c/chromium/src/+/2026147

The problem with the above CL was that the default fill color was set
for all SVGs. This is not according to the spec.

The real problem was that we were reverting fill and stroke colors
for non-SVG elements in forced colors mode. These elements are not
handled in the UA styleshet, and thus will have their fill and stroke
colors set to the default black.

If an SVG element inherits the fill or stroke colors from a non-SVG
parent, it will similarly get set to the default black color.
To fix this, this change avoids reverting fill and stroke for non-
SVG elements to allow for proper inheritance in this case.

Bug: 970285
Change-Id: Ib37cffc5015d01996836d5a22d24fafc9a247522