IndexedDB: Drop max numbers in parallel-cursors-upgrade.html
The external/wpt/IndexedDB/parallel-cursors-upgrade.html test
exercises 10k cursors in parallel, which an implementation should
support. The test was historically flaky, although it hasn't timed out
on the bots lately. But it is still slow - requiring multiple seconds
on many platforms, and 20-40 seconds on debug builds. Drop the maximum
number of parallel cursors to stop being such a drain on the commons.

Local tests: the time for --iterations=40 dropped from 23s to 5s.

Bug: 869364
Change-Id: I5a0b9de3b37650917e40cc0c34e7473a3870b21f